Fixed some inconsistencys between header and cpp function definitions

+++ gl2d.h
@@ -37,7 +37,7 @@ namespace RGNDS {
         void glBegin2D( float screenoffsety = 0 );
         void glEnd2D();
 
-        void glShape(GL_GLBEGIN_ENUM mode, int color, int numPoints, const Point<float> aPoints[], Transform* t = &Transform::_default, int alpha=0, int zDepth=0);
+        void glShape(GL_GLBEGIN_ENUM mode, int color, int numPoints, const Point<double> aPoints[], Transform* t = &Transform::_default, int alpha=0, int zDepth=0);
         void glSprite( int x, int y, int flipmode, const glImage *spr, Transform* tra = &Transform::_default, int zDepth=0);
 
         void glPixel(int x, int y, int color, int alpha=31, int zDepth=0);
